projects/learning-journal/brain-bit-ingestor/requirements.txt
azure-eventhub==1.3.1
environs
pyrebase
tweepy

# non-functional
pytest
